Finish the implementation of routes :tada:
    The routes are implemented and functional, returning appropriate
data and http status to indicate whether the request found a server, if
the request was done incorrectly or if everything was done correctly.

    The message.go and process.go are added in this commit to supply the
routes with specific functionality. A refactor is done in this commit to
make things more readable on the parser package and, as a consequence,
on the main package.

Signed-off-by: Rafael Campos Nunes <rafaelnunes@engineer.com>